Retrofit Express 4/5 dual scheme
Run the integration tests against both Express 4 and Express 5 instead
of the single Express 4 line, mirroring the dual-alias scheme already
established by the rest of the family (express-tee, express-compress,
etc.).

Test layout:
- Move per-topic logic out of `test/<topic>.test.ts` into
  `test/lib/<topic>.ts`, each exporting a `run<Topic>Tests(label,
  express)` runner.
- Add `test/express4.test.ts` and `test/express5.test.ts` that import
  the express factory under the corresponding npm alias and call every
  runner.
- `test/lib/util.ts` follows the family convention with an
  `ExpressFactory = any` shape that satisfies both 4.x and 5.x app
  builders.

Dependencies:
- Drop direct `express` devDep (no `import express from "express"`
  remains in tests after the split).
- Loosen `@types/express` to `^4` so the bare `import type { ... }
  from "express"` references in lib/ still resolve.
- Add the dual-alias quartet `express4` / `express5` /
  `@types/express4` / `@types/express5` for per-version coverage.

Test counts go from 35 (Express 4 only) to 70 (both versions), all
passing.
