+++ package.json
@@ -15,10 +15,13 @@
   "devDependencies": {
     "@rollup/plugin-node-resolve": "^16.0.3",
     "@rollup/plugin-sucrase": "^5.1.0",
-    "@types/express": "^4.17.25",
+    "@types/express": "^4",
+    "@types/express4": "npm:@types/express@^4",
+    "@types/express5": "npm:@types/express@^5",
     "@types/node": "^24.12.2",
     "@types/supertest": "^7.2.0",
-    "express": "^4.21.1",
+    "express4": "npm:express@^4",
+    "express5": "npm:express@^5",
     "middleware-supertest": "^1.1.0",
     "rollup": "^4.60.2",
     "supertest": "^7.2.2",
